requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
alembic
altgraph
annotated-doc
annotated-types
antlr4-python3-runtime
anyio
apache-airflow
apache-airflow-providers-common-compat
apache-airflow-providers-common-io
apache-airflow-providers-common-sql
apache-airflow-providers-fab
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-smtp
apache-airflow-providers-sqlite
apispec
appdirs
argcomplete
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asttokens
async-lru
attrs
babel
beautifulsoup4
bigtree
bleach
blinker
brotli
cached-property
cachelib
cachetools
catboost
certifi
cffi
charset-normalizer
click
clickclick
cloudpickle
colorama
colorlog
comm
configupdater
connexion
contourpy
cron-descriptor
croniter
cryptography
cycler
dask
datasets
datetime
debugpy
decorator
defusedxml
deprecated
dill
dnspython
docker-pycreds
email-validator
et-xmlfile
eval-type-backport
exceptiongroup
executing
fastapi
fastjsonschema
feast
filelock
flask
flask-appbuilder
flask-babel
flask-caching
flask-jwt-extended
flask-limiter
flask-login
flask-session
flask-sqlalchemy
flask-wtf
fonttools
fqdn
frozenlist
fsspec
gitdb
gitpython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-re2
googleapis-common-protos
graphviz
greenlet
grpcio
grpcio-status
gunicorn
h11
h2
hf-xet
hpack
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
hyperframe
idna
importlib-metadata
importlib-resources
inflection
intel-openmp
ipykernel
ipython
ipython-pygments-lexers
isoduration
itsdangerous
jedi
jinja2
jmespath
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-client
jupyter-core
jupyter-events
jupyter-lsp
jupyter-server
jupyter-server-terminals
jupyterlab
jupyterlab-pygments
jupyterlab-server
kiwisolver
lark
lazy-object-proxy
librt
limits
linkify-it-py
locket
lockfile
mako
markdown-it-py
markupsafe
marshmallow
marshmallow-oneofschema
marshmallow-sqlalchemy
matplotlib
matplotlib-inline
mcp
mdit-py-plugins
mdurl
methodtools
mistune
mkl
mmh3
more-itertools
mpmath
multidict
multiprocess
munkres
mypy
mypy-extensions
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook-shim
notion-client
numpy
ollama
omegaconf
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
overrides
packaging
pandas
pandocfilters
parso
partd
pathspec
patsy
pefile
pendulum
pickleshare
pillow
pip
platformdirs
plotly
pluggy
prison
prometheus-client
prompt-toolkit
propcache
proto-plus
protobuf
psutil
pure-eval
pyarrow
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic-core
pydantic-settings
pygments
pyinstaller
pyinstaller-hooks-contrib
pyjwt
pyopenssl
pyparsing
pyside6
pyside6-addons
pyside6-essentials
pysocks
python-daemon
python-dateutil
python-dotenv
python-json-logger
python-multipart
python-nvd3
python-slugify
pytz
pyu2f
pyyaml
pyzmq
referencing
regex
requests
requests-toolbelt
rfc3339-validator
rfc3986-validator
rfc3987-syntax
rich
rich-argparse
rpds-py
rsa
rtdl-revisiting-models
safetensors
scikit-learn
scipy
seaborn
send2trash
sentence-transformers
sentry-sdk
setproctitle
setuptools
shellingham
shiboken6
six
smmap
sniffio
soupsieve
sqlalchemy
sqlalchemy-jsonfield
sqlalchemy-utils
sqlparse
sse-starlette
stack-data
starlette
statsmodels
sympy
tabulate
tbb
tenacity
termcolor
terminado
text-unidecode
threadpoolctl
time-machine
tinycss2
tokenizers
toml
tomli
toolz
torch
torchaudio
torchvision
tornado
tqdm
traitlets
transformers
typeguard
typer
typing-extensions
typing-inspection
typing-utils
tzdata
uc-micro-py
unicodedata2
universal-pathlib
uri-template
uritemplate
urllib3
uvicorn
uvicorn-worker
wandb
watchfiles
wcwidth
webcolors
webencodings
websocket-client
websockets
werkzeug
wheel
win-inet-pton
wirerope
wrapt
wtforms
xxhash
yarl
zipp
zope-interface
zstandard
